+++ symengine/sparse_matrix.cpp
@@ -323,7 +323,8 @@ CSRMatrix CSRMatrix::from_coo(unsigned row, unsigned col,
                               const std::vector<unsigned> &j,
                               const vec_basic &x)
 {
-    unsigned nnz = x.size();
+    // cast is okay, because CSRMatrix indices are unsigned.
+    unsigned nnz = numeric_cast<unsigned>(x.size());
     std::vector<unsigned> p_ = std::vector<unsigned>(row + 1, 0);
     std::vector<unsigned> j_ = std::vector<unsigned>(nnz);
     vec_basic x_ = vec_basic(nnz);
